Move the Employee file to the models folder
