+++ src/main/java/org/example/models/Employee.java
@@ -1,5 +1,31 @@
 package org.example.models;
 
 public class Employee {
-    
+    private int id;
+    private String firstName;
+    private String lastName;
+    private String photo;
+
+    public Employee(int id, String firstName, String lastName, String photo) {
+        this.id = id;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.photo = photo;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public String getPhoto() {
+        return photo;
+    }
 }
